+++ packages/react/src/stories/Basic.stories.tsx
@@ -1,7 +1,8 @@
 import { Meta } from '@storybook/react';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 
 import { Graph, useTrrack } from './useTrrack';
+import { NodeId } from '@trrack/core';
 
 export const LinearExample: React.FC = () => {
     const { trrack: t, actions } = useTrrack();
@@ -142,6 +143,142 @@ export const VertScrolling: React.FC = () => {
 
     return t ? (
         <Graph actions={actions} verticalSpace={120} trrack={t} />
+    ) : (
+        <div></div>
+    );
+};
+
+export const Bookmarks: React.FC = () => {
+    const { trrack: t, actions } = useTrrack();
+
+    const [bookmarkNodes, setBookmarkNodes] = useState<NodeId[]>([]);
+
+    useEffect(() => {
+        t.apply('Add task', actions.addTask({ id: '1', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '2', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '3', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '4', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '5', complete: false }));
+        t.undo();
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '6', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '7', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '8', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '9', complete: false }));
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '14', complete: false }));
+        t.undo();
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '10', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '11', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '12', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '13', complete: false }));
+    }, []);
+
+    return t ? (
+        <Graph
+            actions={actions}
+            isBookmarked={(n: NodeId) => bookmarkNodes.includes(n)}
+            bookmarkNode={(n: NodeId) =>
+                bookmarkNodes.includes(n)
+                    ? setBookmarkNodes(
+                          bookmarkNodes.filter((node) => node !== n)
+                      )
+                    : setBookmarkNodes([...bookmarkNodes, n])
+            }
+            trrack={t}
+        />
+    ) : (
+        <div></div>
+    );
+};
+
+export const Annotations: React.FC = () => {
+    const { trrack: t, actions } = useTrrack();
+
+    const [annotations, setAnnotations] = useState<Record<NodeId, string>>({});
+
+    useEffect(() => {
+        t.apply('Add task', actions.addTask({ id: '1', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '2', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '3', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '4', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '5', complete: false }));
+        t.undo();
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '6', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '7', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '8', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '9', complete: false }));
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '14', complete: false }));
+        t.undo();
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '10', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '11', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '12', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '13', complete: false }));
+    }, []);
+
+    return t ? (
+        <Graph
+            actions={actions}
+            getAnnotation={(n: NodeId) => annotations[n] || ''}
+            annotateNode={(n: NodeId, s: string) =>
+                setAnnotations({ ...annotations, [n]: s })
+            }
+            trrack={t}
+        />
+    ) : (
+        <div></div>
+    );
+};
+
+export const BookmarksAndAnnotations: React.FC = () => {
+    const { trrack: t, actions } = useTrrack();
+
+    const [annotations, setAnnotations] = useState<Record<NodeId, string>>({});
+    const [bookmarkNodes, setBookmarkNodes] = useState<NodeId[]>([]);
+
+    useEffect(() => {
+        t.apply('Add task', actions.addTask({ id: '1', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '2', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '3', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '4', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '5', complete: false }));
+        t.undo();
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '6', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '7', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '8', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '9', complete: false }));
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '14', complete: false }));
+        t.undo();
+        t.undo();
+        t.apply('Add task', actions.addTask({ id: '10', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '11', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '12', complete: false }));
+        t.apply('Add task', actions.addTask({ id: '13', complete: false }));
+    }, []);
+
+    return t ? (
+        <Graph
+            actions={actions}
+            isBookmarked={(n: NodeId) => bookmarkNodes.includes(n)}
+            bookmarkNode={(n: NodeId) =>
+                bookmarkNodes.includes(n)
+                    ? setBookmarkNodes(
+                          bookmarkNodes.filter((node) => node !== n)
+                      )
+                    : setBookmarkNodes([...bookmarkNodes, n])
+            }
+            getAnnotation={(n: NodeId) => annotations[n] || ''}
+            annotateNode={(n: NodeId, s: string) =>
+                setAnnotations({ ...annotations, [n]: s })
+            }
+            trrack={t}
+        />
     ) : (
         <div></div>
     );